refactor: restructure schema files and update Drizzle configuration for improved organization and clarity
