+++ src/drizzle/drizzle.module.ts
@@ -2,7 +2,7 @@ import { Module, Global } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { Pool } from 'pg';
 import { drizzle, NodePgDatabase } from 'drizzle-orm/node-postgres';
-import * as schema from './schema';
+import * as schema from './schema/schema';
 
 // unique identifier for drizzle provider
 export const DRIZZLE = Symbol('drizzle-connection');
