+++ src/drizzle/schema/schema.ts
@@ -1,0 +1,3 @@
+export * from './students.schema';
+export * from './classes.schema';
+export * from './students.schema';
