+++ src/drizzle/schema/users.schema.ts
@@ -1,0 +1,12 @@
+import { timestamp, uuid, varchar, pgTable, pgEnum } from 'drizzle-orm/pg-core';
+
+export const roleEnum = pgEnum('role', ['admin', 'teacher', 'student']);
+
+// Users Table
+export const users = pgTable('users', {
+  id: uuid('id').primaryKey().defaultRandom(),
+  name: varchar('name', { length: 255 }).notNull(),
+  email: varchar('email', { length: 255 }).unique().notNull(),
+  role: roleEnum('role'),
+  createdAt: timestamp('createdAt').defaultNow(),
+});
