--- src/drizzle/types.ts
@@ -1,3 +1,0 @@
-import { pgEnum } from 'drizzle-orm/pg-core';
-
-export const roleEnum = pgEnum('role', ['admin', 'teacher', 'student']);
